+++ library/Episciences/Submit.php
@@ -768,7 +768,7 @@ class Episciences_Submit
     /**
      * @param $repoId
      * @param $id
-     * @param int|null $version
+     * @param float|null $version
      * @param null $latestObsoleteDocId
      * @param bool $manageNewVersionErrors Allow to ignore new version errors for imports
      * @param int|null $rvId
@@ -778,8 +778,8 @@ class Episciences_Submit
      */
     public static function getDoc(
         $repoId,
-        $id,
-        ?int $version = null,
+        &$id,
+        ?float &$version = null,
         $latestObsoleteDocId = null,
         bool $manageNewVersionErrors = true,
         ?int $rvId = null,
@@ -792,11 +792,9 @@ class Episciences_Submit
 
         $isRequiredVersion = true;
 
-        $id = trim($id);
         $isNewVersionOf = !empty($latestObsoleteDocId);
 
         $id = self::cleanIdentifier($id, $repoId);
-
         $translator = Ccsd_Tools::isFromCli() ? null : Zend_Registry::get('Zend_Translate');
         $result = [];
         $oai = null;
@@ -819,7 +817,7 @@ class Episciences_Submit
             }
 
             if (isset($hookVersion['version'])) {
-                $version = $hookVersion['version'];
+                $version = (float)$hookVersion['version'];
                 $result['hookVersion'] = $version;
             }
 
@@ -827,6 +825,9 @@ class Episciences_Submit
             $identifier = Episciences_Repositories::getIdentifier($repoId, $id, $version);
 
             $result['record'] = self::loadRecord($hookApiRecord, $repoId, $identifier, $result, $oai);
+            $result['hookVersion'] = $version;
+            $result['hookId'] = $id;
+            $result['hookRepoId'] = $repoId;
 
             self::mergeEnrichmentsFromHook($hookApiRecord, $result);
 
@@ -852,7 +853,7 @@ class Episciences_Submit
 
             $docId = $paper->alreadyExists();
 
-            self::assertDateTimeVersion($docId, $paper,$result);
+            self::assertDateTimeVersion($docId, $paper, $result);
 
             $result['status'] = $docId ? 2 : 1;
 
@@ -901,7 +902,7 @@ class Episciences_Submit
 
     private static function assertDateTimeVersion(&$docId, Episciences_paper $paper, array &$result): void
     {
-            $currentVersionDateTime = $result[Episciences_Repositories_CryptologyePrint_Hooks::UPDATE_DATETIME] ?? null;
+        $currentVersionDateTime = $result[Episciences_Repositories_CryptologyePrint_Hooks::UPDATE_DATETIME] ?? null;
         if (
             !$docId ||
             empty($currentVersionDateTime)
@@ -925,6 +926,7 @@ class Episciences_Submit
      */
     private static function cleanIdentifier(string $id, $repoId): string
     {
+        $id = trim($id);
         $hookCleanIdentifiers = Episciences_Repositories::callHook('hookCleanIdentifiers', [
             'id' => $id,
             'repoId' => $repoId,
@@ -1563,7 +1565,7 @@ class Episciences_Submit
 
         if (!$canReplace) {
             $suggestedEditors = $this->getSuggestedEditorsFromPost($data);
-            return $this->assignEditors($paper, $suggestedEditors, $data['SID'], $data['VID']);
+            return $this->assignEditors($paper, $suggestedEditors, $data['SID'] ?? null, $data['VID'] ?? null);
         }
 
         return $paper->getEditors(true, true);
@@ -1602,7 +1604,7 @@ class Episciences_Submit
         $oldStatus = (int)Ccsd_Tools::ifsetor($data['old_paper_status'], 0);
         $canReplace = (bool)Ccsd_Tools::ifsetor($data['can_replace'], false);
 
-        if($canReplace){
+        if ($canReplace) {
             $authorTemplateKy = Episciences_Mail_TemplatesManager::TYPE_PAPER_SUBMISSION_UPDATED_AUTHOR_COPY;
         }
 
@@ -2780,7 +2782,7 @@ class Episciences_Submit
 
     }
 
-    public static function  processBasicIdentifier(string &$identifier, $data): void
+    public static function processBasicIdentifier(string &$identifier, $data): void
     {
 
         $identifier = trim($identifier);
@@ -2792,6 +2794,40 @@ class Episciences_Submit
             $identifier .= '/' . $data[Episciences_Repositories_CryptologyePrint_Hooks::UPDATE_DATETIME];
         }
 
+    }
+
+    /**
+     * @param array $post
+     * @return void
+     * to prevent possible data corruption
+     */
+
+    public static function normalizeSubmissionParameters(array &$post): void
+    {
+        // todo Check both forms for the initial submission and new version
+        // In some cases, the hidden field is used; in other cases, it is not.
+
+        $key = 'search_doc';
+
+        $post['h_repoId'] = (int)$post['h_repoId'];
+        $post['h_version'] = (float)$post['h_version'];
+
+        if (isset($post[$key]['docId'])) {
+            $post[$key]['docId'] = $post['h_doc'];
+        }
+
+        if (isset($post[$key]['repoId'])) {
+            $post[$key]['repoId'] = $post['h_repoId'];
+        }
+
+        if (isset($post[$key]['version'])) {
+            $post[$key]['version'] = $post['h_version'];
+        }
+
+        // données vérifiées (nettoyées si nécessaire) lors de la recherche du document
+        $post[$key]['h_docId'] = $post['h_doc'];
+        $post[$key]['h_version'] = $post['h_version'];
+        $post[$key]['h_repoId'] = $post['h_repoId'];
     }
 
 }